Changed filestructure to introduce dumb components and containers
