--- src/DataAccess/Models/EntityBase.cs
@@ -1,11 +1,0 @@
-using System;
-
-namespace DataAccess.Models
-{
-    public class EntityBase
-    {
-        public int Id { get; set; }
-        public DateTime CreatedAt { get; set; }
-        public DateTime LastModified { get; set; }
-    }
-}
