+++ src/WebApi/ClientApp/src/app/app-routing.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { HomeLayoutComponent } from './modules/common/main/layout/home-layout.component';
-import { NotFoundComponent } from './modules/common/notfound/notfound.component';
+import { ContainerComponent } from '@roomsight/container/container.component';
+import { NotFoundComponent } from './shared/notfound/notfound.component';
 
 const routes: Routes = [
-  { path: '', component: HomeLayoutComponent },
+  { path: '', component: ContainerComponent },
   { path: '**', component: NotFoundComponent }
 ];
