+++ src/WebApi/ClientApp/src/app/app.module.ts
@@ -2,9 +2,9 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { ConfigurationModule } from './modules/configurations/configuration.module';
-import { NavbarComponent } from './modules/common/navbar/navbar.component';
-import { RoomsightModule } from './modules/roomsight/roomsight.module';
+import { ConfigurationModule } from './configurations/configuration.module';
+import { NavbarComponent } from './shared/navbar/navbar.component';
+import { RoomsightModule } from './roomsight/roomsight.module';
 
 @NgModule({
   declarations: [
