+++ src/WebApi/ClientApp/src/app/configurations/components/node-card/node-card.component.ts
