+++ src/WebApi/ClientApp/src/app/configurations/components/node-details/node-details.component.html
