+++ src/WebApi/ClientApp/src/app/configurations/configuration.module.ts
@@ -2,11 +2,7 @@ import { NgModule } from "@angular/core";
 import { ConfigurationRoutingModule } from "./configuration-routing.module";
 
 @NgModule({
-  // Declare Components and Directives
-  declarations: [  ],
-  // Modules imported and consumed by app:
   imports: [ ConfigurationRoutingModule ],
-  // Dependencies to inject:
   providers: [],
 })
 export class ConfigurationModule {}
