+++ src/WebApi/ClientApp/src/app/configurations/container/container.component.html
@@ -1,0 +1,2 @@
+<p>layout works!</p>
+
