+++ src/WebApi/ClientApp/src/app/configurations/container/container.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-configurations-container',
+  templateUrl: './container.component.html',
+  styleUrls: ['./container.component.sass']
+})
+export class ContainerComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
