--- src/WebApi/ClientApp/src/app/modules/configurations/components/layout/conf-layout.component.html
@@ -1,3 +1,0 @@
-<p>layout works!</p>
-
-<router-outlet></router-outlet>
