--- src/WebApi/ClientApp/src/app/modules/configurations/components/layout/conf-layout.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-layout',
-  templateUrl: './conf-layout.component.html',
-  styleUrls: ['./conf-layout.component.sass']
-})
-export class ConfLayoutComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
