+++ src/WebApi/ClientApp/src/app/roomsight/components/room/room.component.spec.ts
