+++ src/WebApi/ClientApp/src/app/roomsight/container/container.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-room-container',
+  templateUrl: './container.component.html',
+  styleUrls: ['./container.component.sass']
+})
+export class ContainerComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
