+++ src/WebApi/ClientApp/src/app/roomsight/roomsight-routing.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from "@angular/core";
 import { RouterModule, Routes } from "@angular/router";
-import { LayoutComponent } from "./components/layout/layout.component";
-import { RoomComponent } from "./components/room/room.component";
+import { ContainerComponent } from "@roomsight/container/container.component";
+import { RoomComponent } from "@roomsight//components/room/room.component";
 
 const routes: Routes = [
-  { path: 'roomsight', component: LayoutComponent, pathMatch: 'full' },
+  { path: 'roomsight', component: ContainerComponent, pathMatch: 'full' },
   { path: 'roomsight/:room', component: RoomComponent }
 ];
