+++ src/WebApi/ClientApp/src/app/roomsight/roomsight.module.ts
