+++ src/WebApi/ClientApp/src/app/shared/navbar/navbar.component.html
