+++ src/WebApi/ClientApp/tsconfig.json
@@ -3,6 +3,11 @@
   "compileOnSave": false,
   "compilerOptions": {
     "baseUrl": "./",
+    "paths": { 
+      "@roomsight/*": [ "src/app/roomsight/*" ],
+      "@configurations/*": [ "src/app/configurations/*" ],
+      "@shared/*": [ "src/app/shared/*" ]
+    },
     "outDir": "./dist/out-tsc",
     "sourceMap": true,
     "declaration": false,
